requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.7.1=pypi_0
aiohttp=3.7.4=py37h27cfd23_1
argh=0.26.2=py37_0
argunparse=0.1.2=pypi_0
astroid=2.2.5=py37_0
async-timeout=3.0.1=py37h06a4308_0
attrs=19.1.0=py_0
audioread=2.1.6=py37_0
autopep8=1.4.4=pypi_0
backcall=0.1.0=py_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=openblas
bleach=3.1.0=py_0
blis=0.4.1=pypi_0
boto3=1.17.46=pyhd3eb1b0_0
botocore=1.20.50=pyhd3eb1b0_1
brokenaxes=0.4.2=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h516909a_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=1.0.0=pypi_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.12.3=py37h8022711_0
chardet=3.0.4=py37h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
configparser=5.0.2=pyhd3eb1b0_0
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py_1
cymem=2.0.3=pypi_0
cython=0.29.21=pypi_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.8.0=pyhd3eb1b0_0
dbus=1.13.6=he372182_0
de-core-news-sm=2.3.0=pypi_0
decorator=4.4.0=py_0
defusedxml=0.5.0=py_1
dill=0.3.2=pypi_0
docker-pycreds=0.4.0=py_0
docopt=0.6.2=pypi_0
dtw-python=1.1.6=pypi_0
easydict=1.9=pypi_0
en-core-web-sm=2.3.1=pypi_0
entrypoints=0.3=py37_1000
expat=2.2.5=he1b5a44_1003
fastdtw=0.3.4=pypi_0
ffmpeg=4.1.3=h167e202_0
filelock=3.0.12=pypi_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.0=he983fc9_0
fsspec=2021.7.0=pyhd3eb1b0_0
future=0.17.1=pypi_0
gensim=3.8.0=py37h962f231_0
gettext=0.19.8.1=hc5be6a0_1002
gitdb=4.0.5=pypi_0
gitpython=3.1.9=pypi_0
glib=2.58.3=h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
google-api-core=1.14.3=pypi_0
google-auth=1.6.3=pypi_0
google-cloud-core=1.0.3=pypi_0
google-cloud-speech=1.2.0=pypi_0
google-cloud-storage=1.20.0=pypi_0
google-resumable-media=0.4.1=pypi_0
googleapis-common-protos=1.6.0=pypi_0
gql=0.1.0=py_0
graphql-core=3.1.3=pyhd8ed1ab_0
grpcio=1.22.0=pypi_0
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
h5py=2.9.0=nompi_py37h513d04c_1104
hdf5=1.10.5=nompi_h3c11f04_1100
icu=64.2=he1b5a44_0
idna=2.8=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
imageio-ffmpeg=0.4.2=pypi_0
importlib-metadata=1.7.0=pypi_0
intel-openmp=2019.4=243
iopath=0.1.9=pypi_0
ipykernel=5.1.1=py37h5ca1d4c_0
ipython=7.7.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py_1
isort=4.3.21=py37_0
jedi=0.14.1=py37_0
jinja2=2.10.1=py_0
jiwer=1.3.2=pypi_0
jmespath=0.10.0=py_0
joblib=0.13.2=py_0
jpeg=9c=h14c3975_1001
json5=0.8.5=py_0
jsonschema=3.0.2=py37_0
jupyter-tensorboard=0.2.0=pypi_0
jupyter_client=5.3.1=py_0
jupyter_core=4.4.0=py_0
jupyterlab=1.0.4=py37_0
jupyterlab_server=1.0.0=py_1
kiwisolver=1.1.0=py37hc9558a2_0
krippendorff=0.3.0=pypi_0
lame=3.100=h14c3975_1001
layoutparser=0.2.0=pypi_0
lazy-object-proxy=1.4.1=py37h516909a_0
lexical-diversity=0.1.0=pypi_0
libblas=3.8.0=11_openblas
libcblas=3.8.0=11_openblas
libffi=3.2.1=he1b5a44_1006
libflac=1.3.1=hf484d3e_1002
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=11_openblas
libogg=1.3.2=h14c3975_1001
libopenblas=0.3.6=h6e990d7_6
libpng=1.6.37=hed695b0_0
libprotobuf=3.14.0=h8c45485_0
librosa=0.7.0=py_0
libsndfile=1.0.28=hf484d3e_1000
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h57b8799_1003
libuuid=2.32.1=h14c3975_1000
libvorbis=1.3.5=h14c3975_1001
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=hee79883_2
llvmlite=0.29.0=py37hfd453ef_1
lmdb=1.2.1=pypi_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.8.3=he1b5a44_1001
mad=0.15.1b=he1b5a44_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=py37h14c3975_0
matplotlib=3.1.1=py37_1
matplotlib-base=3.1.1=py37he7580a8_1
mccabe=0.6.1=py_1
mistune=0.8.4=py37h14c3975_1000
mkl=2019.4=243
moviepy=1.0.3=pypi_0
msgpack=1.0.2=pypi_0
multidict=5.1.0=py37h27cfd23_2
murmurhash=1.0.2=pypi_0
natsort=6.2.0=pypi_0
nbconvert=5.5.0=py_0
nbformat=4.4.0=py_1
ncurses=6.1=hf484d3e_1002
nettle=3.4.1=h1bed415_1002
networkx=2.6.2=pyhd3eb1b0_0
ninja=1.9.0=h6bb024c_0
nlp=0.4.0=pypi_0
nltk=3.4.5=py37_0
notebook=6.0.0=py37_0
numba=0.45.1=py37hb3f55d8_0
numexpr=2.7.0=pypi_0
numpy=1.19.0=pypi_0
numpy-base=1.20.3=py37h39b7dee_0
nvidia-ml=7.352.0=py_0
olefile=0.46=py_0
opencv-python=4.1.0.25=pypi_0
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1k=h27cfd23_0
packaging=20.4=pypi_0
pandas=0.25.0=py37hb3f55d8_0
pandoc=2.7.3=0
pandocfilters=1.4.2=py_1
parso=0.5.1=py_0
partd=1.2.0=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
patsy=0.5.1=py_0
pcre=8.41=hf484d3e_1003
pexpect=4.7.0=py37_0
pickle5=0.0.11=pypi_0
pickleshare=0.7.5=py37_1000
pillow=6.1.0=py37h6b7be26_1
pip=19.2.1=py37_0
plac=1.1.3=pypi_0
portalocker=2.3.0=pypi_0
preshed=3.0.2=pypi_0
prettytable=0.7.2=pypi_0
proglog=0.1.9=pypi_0
prometheus_client=0.7.1=py_0
promise=2.3=py37h06a4308_0
prompt_toolkit=2.0.9=py_0
protobuf=3.9.1=pypi_0
psutil=5.8.0=py37h27cfd23_1
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
pyarrow=1.0.0=pypi_0
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycocotools=2.0=pypi_0
pycodestyle=2.5.0=pypi_0
pycparser=2.19=py37_1
pygments=2.4.2=py_0
pylint=2.3.1=py37_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37hcca6a23_2
pyrsistent=0.15.4=py37h516909a_0
pysocks=1.7.1=py37_1
pysoundfile=0.10.2=py_1001
pytesseract=0.3.7=pypi_0
python=3.7.3=h33d41f4_1
python-dateutil=2.8.0=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.0.130_cudnn7.6.3_0
pytorch-nlp=0.5.0=pypi_0
pytorch-pretrained-bert=0.6.2=pypi_0
pytz=2019.2=py_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.2=py37h7b6447c_0
pyzmq=18.0.2=py37h1768529_2
qt=5.9.7=h0c104cb_3
readline=8.0=hf8c457e_0
regex=2020.1.8=pypi_0
requests=2.25.1=pyhd3eb1b0_0
resampy=0.2.1=py_1
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.3.6=pyhd3eb1b0_0
sacremoses=0.0.43=pypi_0
scenedetect=0.5.5=pypi_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.21.3=py37hcdab131_0
scipy=1.3.1=py37h921218d_0
seaborn=0.9.0=py_1
semver=2.13.0=pypi_0
send2trash=1.5.0=py_0
sentencepiece=0.1.86=pypi_0
sentry-sdk=0.20.3=pyh44b312d_0
setuptools=50.3.2=pypi_0
shortuuid=1.0.1=py37h89c1867_4
sip=4.19.8=py37hf484d3e_1000
six=1.15.0=pypi_0
slurmpy=0.0.7=pypi_0
smart_open=5.0.0=pyhd3eb1b0_0
smmap=3.0.4=pypi_0
soupsieve=2.2.1=pyhd3eb1b0_0
sox=14.4.2=h27f72bc_1013
spacy=2.3.2=pypi_0
sqlite=3.29.0=hcee41ef_0
srsly=1.0.2=pypi_0
statsmodels=0.10.1=py37hc1659b7_0
subprocess32=3.5.4=pypi_0
tables=3.6.1=pypi_0
tag-fixer=0.0.4=pypi_0
tb-nightly=1.15.0a20190806=pypi_0
tensorboardx=2.1=pypi_0
terminado=0.8.2=py37_0
testpath=0.4.2=py_1001
textblob=0.15.3=pypi_0
thinc=7.4.1=pypi_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.9=hed695b0_1002
tokenizers=0.7.0=pypi_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.3.0=pypi_0
torchsummary=1.5.1=pypi_0
torchtext=0.6.0=py_1
torchvision=0.5.0=py37_cu100
tornado=6.0.3=py37h516909a_0
tqdm=4.48.2=pypi_0
traitlets=4.3.2=py37_1000
transformers=2.9.0=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.25.6=pypi_0
version-query=1.1.0=pypi_0
wandb=0.10.15=pyh44b312d_0
wasabi=0.7.1=pypi_0
watchdog=0.10.4=pypi_0
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
webrtcvad=2.0.10=pypi_0
webvtt-py=0.4.2=pypi_0
werkzeug=0.15.5=pypi_0
wget=3.2=pypi_0
wheel=0.33.4=py37_0
wrapt=1.11.2=py37h516909a_0
x264=1!152.20180806=h14c3975_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xxhash=2.0.0=pypi_0
xz=5.2.4=h14c3975_1001
yacs=0.1.8=pypi_0
yaml=0.1.7=h96e3832_1
yarl=1.6.3=py37h27cfd23_0
youtube-dl=2021.1.24.1=pypi_0
zeromq=4.3.2=he1b5a44_2
zipp=3.1.0=pypi_0
zlib=1.2.11=h516909a_1005
zstd=1.4.0=h3b9ef0a_0
